Break out xde-test topology creation.
That was a... slightly frustrating exercise in struct and scope drop
order.
