+++ bench/.cargo/config.toml
@@ -1,0 +1,3 @@
+# https://github.com/rust-lang/cargo/issues/3946#issuecomment-973132993
+[env]
+CARGO_WORKSPACE_DIR = { value = "..", relative = true }
