+++ lib/opte-test-utils/src/port_state.rs
@@ -94,7 +94,7 @@ pub fn print_port(port: &Port<VpcNetwork>, vpc_map: &VpcMappings) {
 /// port. It's best not to overthink this, but instead look at some
 /// examples.
 ///
-/// ```
+/// ```text
 /// firewall.rules.in => number of inbound rules in the 'firewall' layer
 /// uft.out => number of outbound UFT entries
 /// epoch => the Port's epoch
@@ -343,7 +343,7 @@ macro_rules! set_na {
 
 /// Set the values of a list of fields and assert.
 ///
-/// ```
+/// ```ignore
 /// set_fields!(pav, "port_state=running, epoch=4, firewall.rules.in=6");
 /// ```
 #[macro_export]
@@ -384,7 +384,7 @@ macro_rules! zero_flows {
 /// `<fields list>` is a `, ` separated list of strings appropriate
 /// for the given instruction.
 ///
-/// ```
+/// ```ignore
 /// update!(g1, ["incr:epoch, firewall.flows.out", "set:port_state=running"])
 /// ```
 #[macro_export]
